Moving asset transforming to its own file.
