+++ rollup.config.js
@@ -25,6 +25,7 @@ import assetPlugin from "./asset-plugin.js";
 import { readFileSync } from "fs";
 import constsPlugin from "./consts-plugin.js";
 import ejsAssetPlugin from "./ejs-asset-plugin.js";
+import assetTransformPlugin from "./asset-transform-plugin.js";
 
 // Delete 'dist'
 require("rimraf").sync("dist");
@@ -78,14 +79,14 @@ export default {
         "./src/assets/space-mono-normal.woff2",
         "./src/assets/space-mono-bold.woff2",
         "./src/assets/favicon.ico"
-      ],
-      transformAssets(asset) {
-        if (asset.fileName.includes("manifest-")) {
-          // Remove name hashing
-          asset.fileName = "manifest.json";
-          // Minify
-          asset.source = JSON.stringify(JSON.parse(asset.source));
-        }
+      ]
+    }),
+    assetTransformPlugin(asset => {
+      if (asset.fileName.includes("manifest-")) {
+        // Remove name hashing
+        asset.fileName = "manifest.json";
+        // Minify
+        asset.source = JSON.stringify(JSON.parse(asset.source));
       }
     }),
     chunkNamePlugin(),
